refactor: extract QrCodeService (#203)
* refactor: extract QrCodeService

* fix: update existing qrcode test

* fix: access modifiers of makeQrCode and makeGoQrCode
