+++ src/server/inversify.config.ts
@@ -28,6 +28,7 @@ import { AuthService } from './services/AuthService'
 import { LogoutController } from './controllers/LogoutController'
 import { UrlManagementService } from './services/UrlManagementService'
 import { UserController } from './controllers/UserController'
+import { QrCodeService } from './services/QrCodeService'
 
 function bindIfUnbound<T>(
   dependencyId: symbol,
@@ -61,6 +62,7 @@ export default () => {
   bindIfUnbound(DependencyIds.logoutController, LogoutController)
   bindIfUnbound(DependencyIds.urlManagementService, UrlManagementService)
   bindIfUnbound(DependencyIds.userController, UserController)
+  bindIfUnbound(DependencyIds.qrCodeService, QrCodeService)
 
   container.bind(DependencyIds.s3Bucket).toConstantValue(s3Bucket)
