+++ src/server/services/interfaces/QrCodeServiceInterface.ts
@@ -1,0 +1,5 @@
+import ImageFormat from '../../../shared/util/image-format'
+
+export interface QrCodeServiceInterface {
+  createGoQrCode: (url: string, format: ImageFormat) => Promise<Buffer>
+}
